löst alla varningar

+++ temporary_modifier.h
@@ -16,7 +16,7 @@
 class Temporary_Modifier : public Moving_Object
 {
 public:
-    Temporary_Modifier(double xpos, double ypos, float scale, double xspeed,
+    Temporary_Modifier(double xpos, double ypos, float scale, float xspeed,
          double speedmodifier, std::string filename);
     ~Temporary_Modifier() = default;
 
@@ -36,7 +36,7 @@ protected:
 class Chalmerist : public Temporary_Modifier
 {
 public: 
-    Chalmerist(double xpos, double ypos, float scale, double xspeed,
+    Chalmerist(double xpos, double ypos, float scale, float xspeed,
          double speedmodifier, std::string filename = "Chalmerist.png");
     ~Chalmerist() = default;
 
@@ -50,7 +50,7 @@ private:
 class Can : public Temporary_Modifier
 {
 public: 
-    Can(double xpos, double ypos, float scale, double xspeed,
+    Can(double xpos, double ypos, float scale, float xspeed,
         double speedmodifier, std::string filename = "eybro.png");
     ~Can() = default;
 
@@ -64,7 +64,7 @@ private:
 class Kir : public Temporary_Modifier
 {
 public: 
-    Kir(double xpos, double ypos, float scale, double xspeed,
+    Kir(double xpos, double ypos, float scale, float xspeed,
         double speedmodifier, std::string filename = "kir.png");
     ~Kir() = default;
